+++ database/migrations/2023_08_21_143803_create_providers_table.php
@@ -24,9 +24,9 @@ return new class extends Migration
             $table->string('instagram')->nullable()->comment("");
             $table->string('description')->comment("");
             $table->string('address')->comment("Direccion fisica del proveedor");
-            $table->integer('stakeholder')->unsigned()->nullable()->comment("Persona de referencia del proveedor");
+            $table->bigInteger('stakeholder')->unsigned()->nullable()->comment("Persona de referencia del proveedor");
             $table->integer('qualification')->comment("Calificacion del proveedor");
-            $table->integer('user_id')->unsigned()->nullable()->comment("Usuario que creó el proveedor");
+            $table->bigInteger('user_id')->unsigned()->nullable()->comment("Usuario que creó el proveedor");
             $table->foreign('user_id')
             ->references('id')
             ->on('users')
