+++ database/migrations/2023_08_21_144139_create_supplierproducts_table.php
@@ -15,9 +15,9 @@ return new class extends Migration
         Schema::create('supplierproducts', function (Blueprint $table) {
             $table->id();
             $table->double('price')->comment("Precio del producto adquirido");
-            $table->integer('product_id')->unsigned()->comment("");
-            $table->integer('provider_id')->unsigned()->comment("");
-            $table->integer('user_id')->unsigned()->nullable()->comment("Usuario que creó el proveedor");
+            $table->bigInteger('product_id')->unsigned()->comment("");
+            $table->bigInteger('provider_id')->unsigned()->comment("");
+            $table->bigInteger('user_id')->unsigned()->nullable()->comment("Usuario que creó el proveedor");
             $table->foreign('user_id')
             ->references('id')
             ->on('users')
